Modernize the minitar codebase.
-   Convert the project to Hoe.
-   Convert tests from test/unit to minitest.
    -   Enable travis-ci tests.
    -   Enable test coverage runs with coveralls.

-   Resolve #3 by using UTC time for a test and test data.
-   Resolve #2 by detecting seekable streams.

-   Deprecate `include Archive::Tar` with a warning.
-   Deprecate `include Archive::Tar::Minitar` with a warning.
-   Add an optional alias `::Minitar` for `Archive::Tar::Minitar`. A
    warning will be printed if `::Minitar` already exists.
-   Moved Archive::Tar::PosixHeader temporarily to
    Archive::Tar::Minitar::PosixHeader. A future release will make it
    ::Minitar::PosixHeader.
-   PosixHeader can now handle Time objects for `mtime` and does not
    require the integer version be provided.
-   Restructured Minitar::Writer.
-   Modified both `Writer#add_file_simple` and `Writer#add_file` to accept a
    `:data` field so that a block does not need to be provided.
