+++ .autotest
@@ -1,0 +1,35 @@
+# -*- ruby -*-
+
+require 'autotest/bundler'
+require 'autotest/restart'
+require 'autotest/timestamp'
+
+def require_plugin(resource)
+  require resource
+rescue LoadError
+  false
+end
+
+require_plugin 'autotest/clear'
+
+Autotest.add_hook :initialize do |at|
+  # at.testlib = "minitest/unit"
+  #
+  # at.extra_files << "../some/external/dependency.rb"
+  #
+  # at.libs << ":../some/external"
+  #
+  # at.add_exception "vendor"
+  #
+  # at.add_mapping(/dependency.rb/) do |f, _|
+  #   at.files_matching(/test_.*rb$/)
+  # end
+  #
+  # %w(TestA TestB).each do |klass|
+  #   at.extra_class_map[klass] = "test/test_misc.rb"
+  # end
+end
+
+# Autotest.add_hook :run_command do |at|
+#   system "rake build"
+# end
