+++ .coveralls.yml
@@ -1,0 +1,2 @@
+---
+service_name: travis-ci
