+++ .gitignore
@@ -1,0 +1,14 @@
+*.md
+*.swp
+*~
+.rake_tasks~
+.source_index
+.vagrant
+Gemfile.lock
+Vagrantfile
+coverage
+doc
+html
+pkg
+publish
+test/cache.tst
