+++ Gemfile
@@ -1,0 +1,14 @@
+# -*- ruby -*-
+
+# NOTE: This file is present to keep Travis CI happy. Edits to it will not
+# be accepted.
+
+source "https://rubygems.org/"
+
+if RUBY_VERSION < "1.9"
+  gem 'mime-types', '~> 1.25'
+end
+
+gemspec
+
+# vim: syntax=ruby
