+++ History.rdoc
@@ -1,0 +1,54 @@
+== 0.6 / YYYY-MM-DD
+
+* Enhancements:
+  * Printing a deprecation warning for including Archive::Tar to put Minitar in
+    the top-level namespace.
+  * Printing a deprecation warning for including Archive::Tar::Minitar into a
+    class (Minitar will be a class for version 1.0).
+  * Moved Archive::Tar::PosixHeader to Archive::Tar::Minitar::PosixHeader with
+    a deprecation warning. Do not depend on Archive::Tar::Minitar::PosixHeader,
+    as it will be moving to ::Minitar::PosixHeader in a near-future release.
+  * Added an alias, ::Minitar, for Archive::Tar::Minitar, opted in with
+    <tt>require 'minitar'</tt>. In future releases, this alias will be enabled
+    by default, and the Archive::Tar namespace will be removed entirely for
+    version 1.0.
+  * Modified the handling of +mtime+ in PosixHeader to do an integer conversion
+    (<tt>#to_i</tt>) so that a Time object can be used instead of the integer value of
+    the time object.
+  * Writer::RestrictedStream was renamed to Writer::WriteOnlyStream for
+    clarity. No alias or deprecation warning was provided for this as it is an
+    internal implementation detail.
+  * Writer::BoundedStream was renamed to Writer::BoundedWriteStream for
+    clarity. A deprecation warning is provided on first use because a
+    BoundedWriteStream may raise a BoundedWriteStream::FileOverflow exception.
+  * Modified <tt>Writer#add_file_simple</tt> to accept the data for a
+    file in <tt>opts[:data]</tt>. When <tt>opts[:data]</tt> is provided, a
+    stream block must not be provided. Improved the documentation for
+    this method.
+  * Modified <tt>Writer#add_file</tt> to accept <tt>opts[:data]</tt> and
+    transparently call <tt>Writer#add_file_simple </tt>in this case.
+
+* Bugs:
+  * Fix {#2}[https://github.com/halostatue/minitar/issues/2] to handle IO
+    streams that are not seekable, such as pipes, STDIN, or STDOUT.
+  * Fix {#3}[https://github.com/halostatue/minitar/issues/3] to make the test
+    timezone resilient.
+
+* Development:
+  * Modernized minitar tooling around Hoe.
+  * Added travis and coveralls.
+
+== 0.5.2 / 2008-02-26
+
+* Bugs:
+  * Fixed a Ruby 1.9 compatibility error.
+
+== 0.5.1 / 2004-09-27
+
+* Bugs:
+  * Fixed a variable name error.
+
+== 0.5.0
+
+* Initial release. Does files and directories. Command does create, extract,
+  and list.
