+++ Licence.rdoc
@@ -1,0 +1,15 @@
+== Licence
+
+* Copyright 2004–2014 Austin Ziegler.
+* Portions copyright 2004 Mauricio Julio Fernández Pradier.
+* Portions copyright 2001–2004 Satoru Takabayashi.
+
+This program is free software. It may be redistributed and/or modified under
+the terms of the GNU GPL version 2 (or, at your option, a later version), or
+Ruby’s licence.
+
+=== GNU GPL version 2
+See the file docs/COPYING.txt in the main distribution.
+
+=== Ruby’s license
+See the file docs/ruby.txt in the main distribution.
