+++ Manifest.txt
@@ -1,0 +1,29 @@
+.autotest
+.gemtest
+.gitignore
+.hoerc
+Contributing.rdoc
+History.rdoc
+Install
+Licence.rdoc
+Manifest.txt
+README.rdoc
+Rakefile
+bin/minitar
+docs/COPYING.txt
+docs/ruby.txt
+lib/archive-tar-minitar.rb
+lib/archive/tar/minitar.rb
+lib/archive/tar/minitar/command.rb
+lib/archive/tar/minitar/input.rb
+lib/archive/tar/minitar/output.rb
+lib/archive/tar/minitar/posix_header.rb
+lib/archive/tar/minitar/reader.rb
+lib/archive/tar/minitar/writer.rb
+lib/minitar.rb
+test/minitest_helper.rb
+test/test_tar_header.rb
+test/test_tar_input.rb
+test/test_tar_output.rb
+test/test_tar_reader.rb
+test/test_tar_writer.rb
