+++ lib/archive-tar-minitar.rb
@@ -1,0 +1,3 @@
+# coding: utf-8
+
+require 'archive/tar/minitar'
